Moved SuperStudent to personal file
