+++ src/main/kotlin/ru/gr106/Student.kt
@@ -1,0 +1,14 @@
+package ru.gr106
+
+class SuperStudent(name:String, marks:Map<String,Int>) : AStudent{
+    override val name: String = name
+    var mmarks = mapOf<String,Int>()
+
+    init {
+        mmarks = marks
+    }
+
+    override fun toString(): String {
+        return "This student is called $name. His marks: ${mmarks.entries}"
+    }
+}
